% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_objective_report.R
\name{aggregate_objective_report}
\alias{aggregate_objective_report}
\title{aggregate_objective_report}
\usage{
aggregate_objective_report(objective_report)
}
\arguments{
\item{objective_report}{casal2_objective_function which is a report from an extract.mpd() object}
}
\value{
data frame of aggregrated objective likelihood components
}
\description{
aggregate_objective_report
}
\details{
take a Casal2 objective_function report and aggregrate components so easier to handle with
visualising likelihood components
}
