% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_objective_report.R
\name{aggregate_single_objective_report}
\alias{aggregate_single_objective_report}
\title{aggregate_single_objective_report}
\usage{
aggregate_single_objective_report(values)
}
\arguments{
\item{values}{data.frame from a Casal2 objective_function report}
}
\value{
data frame with aggregated values
}
\description{
aggregate_single_objective_report
}
\details{
used by aggregate_objective_report
}
