% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_assessment_bookdown.R
\name{build_assessment_bookdown}
\alias{build_assessment_bookdown}
\title{build_assessment_bookdown}
\usage{
build_assessment_bookdown(
  csl_dir,
  output_folder_name,
  mpd_filename,
  config_filename = "config.csl2",
  model_label = NULL,
  verbose = F,
  prompt_user_before_deleting = T
)
}
\arguments{
\item{csl_dir}{directory path to where the Casal2 model is. Currently this function assumes config_filename and mpd_filename are in this directory}

\item{output_folder_name}{foldername for bookdown. It will be relative to csl_dir}

\item{mpd_filename}{the filename for the Casal2 model output}

\item{config_filename}{the filename for the Casal2 config. Default is 'config.csl2' can be replaced with others if you use the -c argument when running Casal2}

\item{model_label}{string a model label used for headers}

\item{verbose}{if the function unexpectedly quits errors, set this to T to see where the problem lies}

\item{prompt_user_before_deleting}{<bool> everytime this function is run it can delete existing bookdowns. Should always be T, unless used in unittests}
}
\value{
will save a suite of Rmd files in output_dir that can be compiled as a bookdown. It should compile the bookdown in the folder _book
}
\description{
This function will build a skeleton/template html bookdown summarizing a single model MPD (casal2 -r, -e)
}
\examples{
\dontrun{
library(r4Casal2)
}
}
\author{
Craig Marsh
}
