% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_report_label.R
\name{check_report_label}
\alias{check_report_label}
\title{Utility function to check reports inputs are consistent}
\usage{
check_report_label(report_label, model)
}
\arguments{
\item{report_label}{string report label}

\item{model}{casal2 model}
}
\description{
Utility function to check reports inputs are consistent
}
\author{
Craig Marsh
}
\keyword{internal}
