% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulation_reports.R
\name{create_simulation_reports}
\alias{create_simulation_reports}
\title{create_simulation_reports}
\usage{
create_simulation_reports(
  config_dir = "",
  config_files,
  output_folder = "simulated_observations",
  prefix = "sim_",
  single_sim_filename = NULL,
  verbose = FALSE
)
}
\arguments{
\item{config_dir}{path directory}

\item{config_files}{vector of string names that correspond to Casal2}

\item{output_folder}{folder directory label to write simulated reports into}

\item{prefix}{prefex for simulated observations}

\item{single_sim_filename}{optional filename, if provided then will report simulated observations into a single file, rather than seperate files.}

\item{verbose}{true print out statements throught function to help debug}
}
\value{
will create an output directory based on parameter output_folder and also a Casal2 report config called simulated_reports.csl2
}
\description{
A function to create a simulated report csl file so that you can add it with an !include statement for simulations.
}
\examples{
\donttest{
library(Casal2)
# plotting Standard Output
config_dir = system.file("extdata", "PosteriorPredictiveChecks", package = "r4Casal2", mustWork = TRUE)
config_files = "Observation.csl2"
obs = create_simulation_reports(config_dir = config_dir, config_files = config_files)
}
}
\author{
Craig Marsh
}
