% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{drop_runs_from_multi_input_mpd}
\alias{drop_runs_from_multi_input_mpd}
\title{drop_runs_from_multi_input_mpd}
\usage{
drop_runs_from_multi_input_mpd(model, elements_to_drop)
}
\arguments{
\item{model}{a model that has been run with -i format}

\item{elements_to_drop}{vector of integers that will be deleted from the model}
}
\value{
model that has been modified
}
\description{
this function will remove runs from an extract.mpd that is a multiinput paramter run
}
