% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_value_table.R
\name{error_value_table}
\alias{error_value_table}
\title{error_value_table}
\usage{
error_value_table(model)
}
\arguments{
\item{model}{<casal2MPD> object that are generated from one of the extract() functions. If list then we expect multiple mpd runs (should be a named list )}
}
\value{
data.frame that can be used by kable or other table functions.
}
\description{
This function will run through all observations and create a table by year and observation summarising type likelihood and error_values
}
\author{
Craig Marsh
}
