% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reweighted_mpds.R
\name{extract_reweighted_mpds}
\alias{extract_reweighted_mpds}
\title{extract_reweighted_mpds}
\usage{
extract_reweighted_mpds(reweighting_path)
}
\arguments{
\item{reweighting_path}{path of the directory that contains the estimated output}
}
\value{
a names list of mpd that can be accessed with the get_ and plot_ methods.
}
\description{
this function will read in a set of re-weighted models as list so you can
plot and integreate the effect of data-weightings. This expects model outputs produced
by the function run_automatic_reweighting
}
\details{
this function expects there to be a range of outputed files with the following filename structure "estimate_'weighting iteration'.log,
where 'weighting iteration' is an integer.
}
