% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.starting.pars.R
\name{generate.starting.pars}
\alias{generate.starting.pars}
\title{generate.starting.pars Generates a parameter file that is formatted for -i input into Casal2.}
\usage{
generate.starting.pars(
  MPD = MPD,
  n_sim = 10,
  start_free_filename = "start_free.dat",
  path = "",
  all_uniform = FALSE,
  quiet = T,
  write = TRUE
)
}
\arguments{
\item{MPD}{<casal2MPD> object that are generated from one of the extract() functions. If list then we expect multiple MPD runs (should be a named list)}

\item{n_sim}{the number of randomly generated sets of free parameters}

\item{start_free_filename}{= the filename of the file created}

\item{path}{Optionally, the path to location of the output file}

\item{all_uniform}{= logical if TRUE draw from a uniform between bounds regardless of the prior distribution}

\item{quiet}{Optional, suppresses printing statements}
}
\value{
a file (with default name 'start_free.dat') in the path directory
}
\description{
This function takes an  Casal2 MPD  object with an estimate_summary report, and returns a free parameters file. Where each parameter is drawn from the prior defined in an @estimate block.
}
\author{
Alistair Dunn
}
