% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_starting_pars.R
\name{generate_starting_pars}
\alias{generate_starting_pars}
\title{generate.starting.pars Generates a parameter file that is formatted for -i input into Casal2.}
\usage{
generate_starting_pars(
  MPD = MPD,
  n_sim = 10,
  start_free_filename = "start_free.dat",
  path = "",
  all_uniform = FALSE,
  shrink = 0,
  quiet = T,
  write = TRUE
)
}
\arguments{
\item{MPD}{<casal2MPD> object that are generated from the extract() functions.}

\item{n_sim}{the number of randomly generated sets of free parameters}

\item{start_free_filename}{= the filename of the file created}

\item{path}{Optionally, the path to location of the output file}

\item{all_uniform}{= logical if TRUE draw from a uniform between bounds regardless of the prior distribution}

\item{shrink}{shrink values near bounds by shrink/2 percent of the range of bounds to keep starting pars away from bounds (default = 0) and range (0.0-0.99)}

\item{quiet}{Optional, suppresses printing statements}

\item{write}{if false, do not write to the output file}
}
\value{
a named matrix of starting parameters, and if write = TRUE then outputs a file containing the starting parameters in the format of a -i file for Casal2
}
\description{
This function takes an  Casal2 MPD  object with an estimate_summary report, and returns a free parameters file.
Each parameter is drawn from the prior defined in an @estimate block and the associated bounds.
Estimates using transformed parameters are not fully implemented.
}
\author{
A Dunn
}
