% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_get_methods.R
\name{get_abundance_observations}
\alias{get_abundance_observations}
\alias{get_abundance_observations.casal2MPD}
\title{get_abundance_observations}
\usage{
get_abundance_observations(model)

\method{get_abundance_observations}{casal2MPD}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract() functions. If list then we expect multiple mpd runs (should be a named list )}
}
\value{
dataframe with all observations of type == 'observation' and observation_type %in% c('biomass', 'abundance')
}
\description{
A plotting function to plot fishing presuure (U's ) for 'casal2TAB' and 'casal2MPD' objects.
}
\author{
Craig Marsh
}
