% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_aggregated_composition_values}
\alias{get_aggregated_composition_values}
\alias{get_aggregated_composition_values.casal2MPD}
\alias{get_aggregated_composition_values.list}
\title{get_aggregated_composition_values}
\usage{
get_aggregated_composition_values(model)

\method{get_aggregated_composition_values}{casal2MPD}(model)

\method{get_aggregated_composition_values}{list}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract() functions. If list then we expect multiple mpd runs (should be a named list )}
}
\value{
data frame with aggregated observed and expected over all years for each observation
}
\description{
An accessor function that returns a data frame of aggregrated composition
}
\author{
Craig Marsh
}
