% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{get_block}
\alias{get_block}
\title{get_block}
\usage{
get_block(parameter_label)
}
\arguments{
\item{parameter_label}{full parameter label 'block[label].parameter'}
}
\value{
the block label
}
\description{
get the block type for a parameter
}
