% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchabilities.R
\name{get_catchabilities}
\alias{get_catchabilities}
\alias{get_catchabilities.casal2MPD}
\alias{get_catchabilities.list}
\title{get_catchabilities}
\usage{
get_catchabilities(model)

\method{get_catchabilities}{casal2MPD}(model)

\method{get_catchabilities}{list}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all derived quantitiy reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of catchabilities
}
\author{
Craig Marsh
}
