% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_composition_mean_bin}
\alias{get_composition_mean_bin}
\alias{get_composition_mean_bin.casal2MPD}
\title{get_composition_mean_bin}
\usage{
get_composition_mean_bin(model, ignore_plus_group = FALSE)

\method{get_composition_mean_bin}{casal2MPD}(model, ignore_plus_group = FALSE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract() functions. If list then we expect multiple mpd runs (should be a named list )}

\item{ignore_plus_group}{<bool> we will assume the max age or length is a plus group and remove it.}
}
\value{
dataframe with all mean age or length and standard errors
}
\description{
An accessor function that returns a data frame of mean age or length calculations
}
\details{
the dataframe returned has Oy: mean age/length observed, Ey: mean age/length predicted, SEy: Standard error, Nassumed: mean effective sample size.
Why use ignore_plus_group? not sure we need to chat to Jeremy and C Francis taken from the SNA code. My guess is that if there is a large plus group length or age, then this can skew the mean
statistic. At first thoughts I don't think it should be employed that frequently
}
\author{
Craig Marsh
}
