% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_derived_quanitites.R
\name{get_derived_quanitites}
\alias{get_derived_quanitites}
\alias{get_derived_quanitites.casal2MPD}
\alias{get_derived_quanitites.list}
\alias{get_derived_quanitites.casal2TAB}
\title{get_derived_quanitites}
\usage{
get_derived_quanitites(model)

\method{get_derived_quanitites}{casal2MPD}(model)

\method{get_derived_quanitites}{list}(model)

\method{get_derived_quanitites}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all derived quantitiy reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of derived_quantities
}
\author{
Craig Marsh
}
