% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dqs.R
\name{get_dqs}
\alias{get_dqs}
\alias{get_dqs.casal2MPD}
\alias{get_dqs.list}
\alias{get_dqs.casal2TAB}
\title{get_dqs}
\usage{
get_dqs(model)

\method{get_dqs}{casal2MPD}(model)

\method{get_dqs}{list}(model)

\method{get_dqs}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all derived quantitiy reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of selectivities
}
\author{
Craig Marsh
}
