% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimated_values.R
\name{get_estimated_values}
\alias{get_estimated_values}
\alias{get_estimated_values.casal2MPD}
\alias{get_estimated_values.list}
\title{get_estimated_values}
\usage{
get_estimated_values(model)

\method{get_estimated_values}{casal2MPD}(model)

\method{get_estimated_values}{list}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all estimate_value reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of estimate_value
}
\author{
Craig Marsh
}
