% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fishing_pressure.R
\name{get_fishing_pressure}
\alias{get_fishing_pressure}
\alias{get_fishing_pressure.casal2MPD}
\alias{get_fishing_pressure.list}
\alias{get_fishing_pressure.casal2TAB}
\title{get_fishing_pressure}
\usage{
get_fishing_pressure(model)

\method{get_fishing_pressure}{casal2MPD}(model)

\method{get_fishing_pressure}{list}(model)

\method{get_fishing_pressure}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of process type recruitment
}
\author{
Craig Marsh
}
