% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_high_correlations.R
\name{get_high_correlations}
\alias{get_high_correlations}
\title{get_high_correlations}
\usage{
get_high_correlations(correlation_matrix, max_correlation = 0.8, labels = NULL)
}
\arguments{
\item{correlation_matrix}{symetric correlation matrix}

\item{max_correlation}{the max correlation value. must be a value between (0 - 1)}

\item{labels}{parameter labels for the correlation matrix}
}
\value{
data frame of row and column index and paramater label if labels supplied along with the correlation value
}
\description{
get highly correlated parameters from a correlation matrix
}
