% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init_F.R
\name{get_init_F}
\alias{get_init_F}
\alias{get_init_F.casal2MPD}
\alias{get_init_F.list}
\alias{get_init_F.casal2TAB}
\title{get_init_F}
\usage{
get_init_F(model, ...)

\method{get_init_F}{casal2MPD}(model, reformat_labels = TRUE)

\method{get_init_F}{list}(model, reformat_labels = TRUE)

\method{get_init_F}{casal2TAB}(model, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of process type mortality_initialisation_baranov
}
\author{
Craig Marsh
}
