% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inital_partition.R
\name{get_initial_partition}
\alias{get_initial_partition}
\alias{get_initial_partition.casal2MPD}
\alias{get_initial_partition.list}
\alias{get_initial_partition.casal2TAB}
\title{get_initial_partition}
\usage{
get_initial_partition(model, ...)

\method{get_initial_partition}{casal2MPD}(model, reformat_labels = TRUE)

\method{get_initial_partition}{list}(model, reformat_labels = TRUE)

\method{get_initial_partition}{casal2TAB}(model, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of process type initialisation_partition
}
\author{
Craig Marsh
}
