% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{get_label}
\alias{get_label}
\title{get_label}
\usage{
get_label(parameter_label)
}
\arguments{
\item{parameter_label}{full parameter label 'block[label].parameter'}
}
\value{
the block label
}
\description{
get the label type for a parameter
}
