% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objective_function.R
\name{get_objective_function}
\alias{get_objective_function}
\alias{get_objective_function.casal2MPD}
\alias{get_objective_function.list}
\title{get_objective_function}
\usage{
get_objective_function(model, aggregate_obs = T)

\method{get_objective_function}{casal2MPD}(model, aggregate_obs = T)

\method{get_objective_function}{list}(model)
}
\arguments{
\item{model}{casal2 model or a list of model}

\item{aggregate_obs}{<bool> whether to aggregate over observations, if false will report objective function by year and observation.}
}
\value{
data frame of objective function negative log-likelihood components
}
\description{
get_objective_function
}
\details{
take a Casal2 objective_function report and aggregate components so easier to handle with
visualising likelihood components
}
