% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{get_parameter}
\alias{get_parameter}
\title{get_parameter}
\usage{
get_parameter(parameter_label)
}
\arguments{
\item{parameter_label}{full parameter label 'block[label].parameter'}
}
\value{
the block label
}
\description{
get the parameter label
}
