% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_process.R
\name{get_process}
\alias{get_process}
\alias{get_process.casal2MPD}
\title{get_process}
\usage{
get_process(model, label, ...)

\method{get_process}{casal2MPD}(model, label, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of a nameed process
}
\author{
Alistair Dunn
}
