% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selectivities.R
\name{get_selectivities}
\alias{get_selectivities}
\alias{get_selectivities.casal2MPD}
\alias{get_selectivities.list}
\alias{get_selectivities.casal2TAB}
\title{get_selectivities}
\usage{
get_selectivities(model)

\method{get_selectivities}{casal2MPD}(model)

\method{get_selectivities}{list}(model)

\method{get_selectivities}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all selectivity reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of selectivities
}
\author{
Craig Marsh
}
