% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selectivities_by_year.R
\name{get_selectivities_by_year}
\alias{get_selectivities_by_year}
\alias{get_selectivities_by_year.casal2MPD}
\alias{get_selectivities_by_year.list}
\title{get_selectivities_by_year}
\usage{
get_selectivities_by_year(model)

\method{get_selectivities_by_year}{casal2MPD}(model)

\method{get_selectivities_by_year}{list}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all selectivity reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of selectivities
}
\author{
Craig Marsh
}
