% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulated_age_resids.R
\name{get_simulated_age_resids}
\alias{get_simulated_age_resids}
\title{get_simulated_age_resids}
\usage{
get_simulated_age_resids(simulated_obs, observation_report)
}
\arguments{
\item{simulated_obs}{<list> output from read.simulated.data() function}

\item{observation_report}{<data.frame> values data frame from casal2 report type = observation}
}
\value{
A list of objects. 
\itemize{
  \item mpd_df data frame with observed expected and quantile residuals (assumed U(0,1)) and normal map residuals (assumed N(0,1))
  \item full_simulated_values data frame with all simulated observations for years and ages
  \item auto_corr_p_val test for significant autocorrelation along age-bins for each year see ?testTemporalAutocorrelation for more information
  \item dispersion_p_val tests for under and over dispersion for each year see ?testDispersion
  \item zero_inflat_p_val tests for zero inflation in each year see ?testZeroInflation
}
}
\description{
A function for calculating simulation-based scaled (quantile) residuals from simulated observations from Casal2 and 
DHARMa inbuilt functionality
}
\details{
this package takes simualted data read into R from r4Casal2::read.simulated.data() function along with MPD observation report estimates
and uses DHARMa (vignette("DHARMa", package="DHARMa")) to calculate simulation-based scaled (quantile) residuals
}
\examples{
\donttest{

}
}
\author{
Craig Marsh
}
