% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timevarying_parameters.R
\name{get_timevarying_parameters}
\alias{get_timevarying_parameters}
\alias{get_timevarying_parameters.casal2MPD}
\alias{get_timevarying_parameters.list}
\title{get_timevarying_parameters}
\usage{
get_timevarying_parameters(model)

\method{get_timevarying_parameters}{casal2MPD}(model)

\method{get_timevarying_parameters}{list}(model)
}
\arguments{
\item{model}{<casal2MPD> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with profile_values and likelihood components
}
\description{
An accessor function that returns a data frame of a time_varying class that can be easily plotted
}
\author{
Craig Marsh
}
