% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformed_parameters.R
\name{get_transformed_parameters}
\alias{get_transformed_parameters}
\alias{get_transformed_parameters.casal2MPD}
\alias{get_transformed_parameters.list}
\alias{get_transformed_parameters.casal2TAB}
\title{get_transformed_parameters}
\usage{
get_transformed_parameters(model)

\method{get_transformed_parameters}{casal2MPD}(model)

\method{get_transformed_parameters}{list}(model)

\method{get_transformed_parameters}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with profile_values and likelihood components
}
\description{
An accessor function that returns a data frame of a parameter_transformations class that can be easily plotted
}
\author{
Craig Marsh
}
