% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abundance_observations.R
\name{has_abundance_observations}
\alias{has_abundance_observations}
\title{has_abundance_observations}
\usage{
has_abundance_observations(model, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions}
}
\value{
TRUE if the object contains abundance_observations
}
\description{
An accessor function that returns if abundance data is in a model
}
\author{
Alistair Dunn
}
