% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchabilities.R
\name{has_catchabilities}
\alias{has_catchabilities}
\title{has_catchabilities}
\usage{
has_catchabilities(model, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions}
}
\value{
TRUE if the object contains catchabilities
}
\description{
An accessor function that returns if catchabilities is in a model
}
\author{
Alistair Dunn
}
