% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_composition_observations.R
\name{has_composition_observations}
\alias{has_composition_observations}
\title{has_composition_observations}
\usage{
has_composition_observations(model, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions}
}
\value{
TRUE if the object contains composition_observations
}
\description{
An accessor function that returns if composition data is in a model
}
\author{
Alistair Dunn
}
