% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.derived_quantities.R
\name{plot.derived_quantities}
\alias{plot.derived_quantities}
\alias{plot.derived_quantities.casal2MPD}
\alias{plot.derived_quantities.casal2TAB}
\alias{plot.derived_quantities.list}
\title{plot.derived_quantities default}
\usage{
plot.derived_quantities(
  model,
  report_label = "",
  plot_type = "classic",
  plot.it = T
)

\method{plot.derived_quantities}{casal2MPD}(
  model,
  report_label = "",
  plot_type = "classic",
  plot.it = T
)

\method{plot.derived_quantities}{casal2TAB}(
  model,
  report_label = "",
  plot_type = "classic",
  plot.it = T
)

\method{plot.derived_quantities}{list}(
  model,
  report_label = "",
  plot_type = "classic",
  plot.it = T
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract() functions. If list then we expect multiple mpd runs (should be a named list )}

\item{plot_type}{string}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}

\item{report_labels}{vector<string> of report labels to plot}
}
\value{
A ggplot object
}
\description{
A plotting function for Casal2 derived_quantities
}
\examples{
\donttest{
library(Casal2)
# plotting Standard Output
data <- extract.mpd(file = system.file("extdata", "estimate.out", package="r4Casal2"))
names(data)
par(mfrow = c(1,2))
plot.derived_quantities(model = data, report_labels = "SSB")
}
}
\author{
Craig Marsh
}
