% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.selectivities.R
\name{plot.selectivities}
\alias{plot.selectivities}
\alias{plot.selectivities.casal2MPD}
\title{plot.selectivities plot selectivities from a Casal2 model.}
\usage{
plot.selectivities(model, report_labels, plot.it = T)

\method{plot.selectivities}{casal2MPD}(model, report_labels, plot.it = T)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}

\item{report_labels}{vector<string>}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}
}
\value{
generate a plot by age or length or if plot.it = F then returns a data.frame

\code{NULL}
}
\description{
A plotting function to plot selectivities for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
Craig Marsh
}
