% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_composition_predictive_dist.R
\name{plot_composition_predictive_dist}
\alias{plot_composition_predictive_dist}
\title{plot_composition_predictive_dist
Plots a violin plot of the predictive distribution by age or length for an observation}
\usage{
plot_composition_predictive_dist(sim_data, obs, lab, plot_type = "violin")
}
\arguments{
\item{sim_data}{matrix of simulated data rows = year, col = simulations. This function assumes rownames(sim_data) = year}

\item{obs}{a data frame with columns obs and year}

\item{lab}{a label for the plot which clearly identifies the observation}

\item{plot_type}{= "violin" currently either violin or boxplot.}

\item{age}{<bool> if true plot ages if FALSE it is length observations}
}
\value{
a ggplot
}
\description{
plot_composition_predictive_dist
Plots a violin plot of the predictive distribution by age or length for an observation
}
