% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compositional_observations.R
\name{plot_compositional_observations}
\alias{plot_compositional_observations}
\alias{plot_compositional_observations.casal2MPD}
\title{plot_compositional_observations}
\usage{
plot_compositional_observations(model, ...)

\method{plot_compositional_observations}{casal2MPD}(model, report_labels = NULL, plot.it = T)
}
\arguments{
\item{model}{<casal2MPD> object that are generated from the extract_mpd() function}

\item{report_labels}{vector<string> report labels to plot, default is all}

\item{plot.it}{Whether to generate a default plot or just return the values as a matrix}

\item{plot_type}{string
\itemize{
  \item classic
  \item residual
}}
}
\value{
A ggplot object

\code{NULL}
}
\description{
A plotting function for Casal2 fits to age and length compositional data
}
\examples{
\dontrun{

}
}
\author{
Craig Marsh
}
