% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{plot_profile}
\usage{
plot_profile(
  profile,
  mpd = NULL,
  objective_function_components = "all",
  ask_for_obj_labels = FALSE,
  likelihood_cut_off = 100,
  plot_style = "individual",
  aggregate_obs = TRUE
)
}
\arguments{
\item{profile}{the profile Casal2 output read in from extract.mpd()}

\item{mpd}{the mpd Casal2 output read in by extract.mpd()}

\item{objective_function_components}{either specify 'all' which will plot all likelihood contributions, or a vector of characters for the contirbutions you want. If not sure how to specify these see ask_for_obj_labels}

\item{ask_for_obj_labels}{boolean if True the function will print the values available for objective_function_components}

\item{likelihood_cut_off}{ylim for likelihood to remove large values obscuring the trend of the profile}

\item{plot_style}{a character specifying the plot type
 \itemize{
  \item individual: separate plots for each component, with filled lines underneath
  \item classic: All components on a single plot with lines.
}}

\item{aggregate_obs}{bool, whether aggregate_obs observations log-likelihoods over all years}
}
\description{
plot_profile
}
\author{
C Marsh
}
