% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\alias{plot_recruitment.casal2MPD}
\alias{plot_recruitment.casal2TAB}
\title{plot_recruitment plot recuitment}
\usage{
plot_recruitment(
  model,
  report_label = "",
  quantity = "ycs_values",
  plot.it = T,
  ...
)

\method{plot_recruitment}{casal2MPD}(
  model,
  report_label = "",
  quantity = "ycs_values",
  plot.it = T,
  ...
)

\method{plot_recruitment}{casal2TAB}(
  model,
  report_label = "",
  quantity = "ycs_values",
  plot.it = T,
  ...
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions using the Casal2 base library}

\item{report_label}{<string>}

\item{quantity}{\itemize{
  \item ycs_values
  \item Recruits
  \item true_ycs
  \item standardised_ycs
}}

\item{plot.it}{Whether to generate a default plot or return the values as a dataframe for personal plots}

\item{...}{remaining plotting options}
}
\value{
generate a plot over time if plot.it = T, if plot.it = F it will return a matrix of values.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to plot recuitment for 'casal2TAB' and 'casal2MPD' objects.
}
\details{
If you have multiple time-steps and Recruitment
}
\author{
Craig Marsh
}
