% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_index.R
\name{plot_relative_index}
\alias{plot_relative_index}
\alias{plot_relative_index.casal2MPD}
\alias{plot_relative_index.casal2TAB}
\title{plot_relative_index default}
\usage{
plot_relative_index(
  model,
  report_labels = NULL,
  plot_type = "classic",
  plot.it = T
)

\method{plot_relative_index}{casal2MPD}(
  model,
  report_labels = NULL,
  plot_type = "classic",
  plot.it = T
)

\method{plot_relative_index}{casal2TAB}(
  model,
  report_labels = NULL,
  plot_type = "classic",
  plot.it = T
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions.}

\item{report_labels}{vector<string> report labels to plot, default is all}

\item{plot_type}{string
\itemize{
  \item classic
  \item classic_ribbon
  \item residual
}}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}
}
\value{
A ggplot object

\code{NULL}

\code{NULL}
}
\description{
A plotting function for Casal2 fits to abundance or biomass observations
}
\examples{
\dontrun{
library(casal2)
# plotting Standard Output
data <- extract.mpd(file = system.file("extdata", "estimate.out", package="casal2"))
names(data)
par(mfrow = c(1,2))
plot.fits(model = data, report_labels = "westF_at_age")
plot.fits(model = data, report_labels = "eastF_at_age")
# to create a user-specified plot, use plot.it = FALSE.
Tangaroa_fits <- plot.fits(model = data, report_label = "eastF_at_age", plot.it = FALSE)
}
}
\author{
Craig Marsh
}
