% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.selectivities.R, R/plot_selectivities.R
\name{plot_selectivities}
\alias{plot_selectivities}
\title{plot_selectivities}
\usage{
plot_selectivities(selectivity_df)

plot_selectivities(selectivity_df)
}
\arguments{
\item{selectivity_df}{data.frame}
}
\value{
a ggplot

a ggplot
}
\description{
A Generic plotting function for selectivities that are derived from get_selectivities function()

A Generic plotting function for selectivities that are derived from get_selectivities function()
}
\author{
Craig Marsh

Craig Marsh
}
