% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.simulated.data.R
\name{read.simulated.data}
\alias{read.simulated.data}
\title{read.simulated.data}
\usage{
read.simulated.data(dir, verbose = FALSE)
}
\arguments{
\item{dir}{path directory to a folder containing simulated observation from a casal2 -s run}

\item{verbose}{true print out statements throught function to help debug}
}
\value{
a list of all the simulated data sets
}
\description{
A utility function for reading in Casal2 simualted data sets. For help on simulated data from Casal2 see
}
\author{
Craig Marsh
}
