% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_automatic_reweighting.R
\name{run_automatic_reweighting}
\alias{run_automatic_reweighting}
\title{run_automatic_reweighting}
\usage{
run_automatic_reweighting(
  config_dir,
  config_filename = "config.csl2",
  weighting_folder_name = "Reweight",
  mpd_file_name = "estimate.log",
  n_loops = 3,
  observation_labels = NULL,
  dash_i_par_filename = NULL,
  prompt_user_before_deleting = TRUE,
  verbose = T
)
}
\arguments{
\item{config_dir}{directory that contains config. See details for info on nested folder configs}

\item{config_filename}{the config file that describes all Casal2 input files. Expected to be in 'config_dir'.}

\item{weighting_folder_name}{A folder name created in 'config_dir' where re-estiamtion is done and output saved.}

\item{mpd_file_name}{filename with mpd output to start calculating weights for. Expected to be in 'config_dir'.}

\item{n_loops}{number of iterative loops}

\item{observation_labels}{if you only want to weight a subset of compositional data sets}

\item{dash_i_par_filename}{parameter filename compatible with -i format. Useful to start models close to solution for complex models}

\item{prompt_user_before_deleting}{if FALSE will delete previous weighting_folder_name without prompting user.}

\item{verbose}{print additional information to screen}
}
\value{
data frame of weights in each loop
}
\description{
a function that will automatically conduct an iterative reweighting re-esitmation.
}
\details{
Sometimes users may have subdirectories containing config files. This function is untested for this config model structure
}
