% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunsTest.R
\name{runs_test_residuals}
\alias{runs_test_residuals}
\title{runs test}
\usage{
runs_test_residuals(x, test = "two.sided")
}
\arguments{
\item{x}{a time-series of residuals}

\item{test}{c("less","greater","two.sided"). Default less is checking for postive autocorrelation only}
}
\value{
runs p value and upper and lower control limits
}
\description{
Run Wald-Wolfowitz Runs Test and calculate upper and lower control limits based on individual/moving-range chart see https://en.wikipedia.org/wiki/Shewhart_individuals_control_chart
}
\details{
for mean age or length residuals we suggest you use the residuals from the from the logged values.
H0: The order of the data is random
HA|two.sided: The order of the data is not random
HA|right.sided: a trend
HA|greater: a first order negative serial correlation
}
