% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{string.to.vector.of.words}
\alias{string.to.vector.of.words}
\title{string.to.vector.of.words}
\usage{
string.to.vector.of.words(string)
}
\value{
vector of words
}
\description{
string.to.vector.of.words
}
