% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_config.R
\name{summarise_config}
\alias{summarise_config}
\title{Summarise config file}
\usage{
summarise_config(config_dir = "", config_file = "config.csl2", quiet = T)
}
\arguments{
\item{config_dir}{Directory for the config file.}

\item{config_file}{Name of config file.}

\item{quiet}{Whether to print information to console.}
}
\value{
A list of summaries from the Casal2 config file.
}
\description{
Summarise config file
}
\examples{
summary = summarise_config(config_dir = 'inst/testdata/Simple/', config_file = "config_betadiff.csl2", quiet = T)


}
