% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_model.R
\name{summarise_model}
\alias{summarise_model}
\title{summarise_model}
\usage{
summarise_model(
  config_dir = "",
  config_file = "config.csl2",
  quiet = T,
  fileEncoding = ""
)
}
\arguments{
\item{config_dir}{path directory}

\item{config_file}{the starting config that starts a Casal2 model default config.csl2, but can be overridden with casal2 -c my_config. in the later case it should be the my_config name}
}
\value{

}
\description{
utility function that reads in config.csl2 file and summarises a model in table formats and plots
}
\author{
Craig Marsh
}
