% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{z_scale}
\alias{z_scale}
\title{Convergence and efficiency diagnostics for Markov Chains These functions are
improved versions of the traditional Rhat (for convergence) and Effective Sample Size (for efficiency).
Credit for this code https://github.com/stan-dev/rstan/blob/develop/rstan/rstan/R/monitor.R
Protected under GPLv3
must remain free and available
Rank normalization}
\usage{
z_scale(x)
}
\arguments{
\item{x}{A numeric array of values.}
}
\value{
A numeric array of rank normalized values with the same
    size as input.
}
\description{
Compute rank normalization for a numeric array. First replace each
value by its rank. Average rank for ties are used to conserve the
number of unique values of discrete quantities. Second, normalize
ranks via the inverse normal transformation.
}
