% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_composition_stage_two_weights.R
\name{calculate_composition_stage_two_weights}
\alias{calculate_composition_stage_two_weights}
\title{calculate_composition_stage_two_weights}
\usage{
calculate_composition_stage_two_weights(
  model,
  observation_labels = NULL,
  approximate_single_year_obs = FALSE
)
}
\arguments{
\item{model}{casal2 mpd model}

\item{observation_labels}{vector<string> of observation labels to calculate the stage two weights for. This function assumes your observation label (@observation) is the same as the report label (@report)}

\item{approximate_single_year_obs}{<bool> see details}
}
\value{
data.frame of second stage weights
}
\description{
this function will iterate over all (or a subset of) compositional observations to calculate the method TA 1.8 weights from \insertCite{francis2011data}{r4Casal2}
}
\details{
if approximate_single_year_obs then we will search across other observations and years to find a an error value similar in magnitude for the year with a single observation
and assign a weight based on the year that is similar in magnitude.
}
\references{
\insertAllCited{}
}
