% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recruitment.R
\name{get_BH_recruitment}
\alias{get_BH_recruitment}
\alias{get_BH_recruitment.casal2MPD}
\alias{get_BH_recruitment.list}
\alias{get_BH_recruitment.casal2TAB}
\title{get_BH_recruitment}
\usage{
get_BH_recruitment(model)

\method{get_BH_recruitment}{casal2MPD}(model)

\method{get_BH_recruitment}{list}(model)

\method{get_BH_recruitment}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame from Casal2 model output

A list frame from Casal2 model output. There is a non_multi_column_df which contains non year varying components such as B0, steepness etc and a multi_column_df which has ssb, recrutis, ycs etc
}
\description{
An accessor function that returns a data frame from a Casal2 model output of process type recruitment
}
\author{
Craig Marsh
}
