% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchabilities.R
\name{get_catchabilities}
\alias{get_catchabilities}
\alias{get_catchabilities.casal2MPD}
\alias{get_catchabilities.list}
\alias{get_catchabilities.casal2TAB}
\title{get_catchabilities}
\usage{
get_catchabilities(model, ...)

\method{get_catchabilities}{casal2MPD}(model, reformat_labels = TRUE)

\method{get_catchabilities}{list}(model, reformat_labels = TRUE)

\method{get_catchabilities}{casal2TAB}(model, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
A data frame with all derived quantity reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of catchabilities
}
\author{
Craig Marsh
}
