% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_transitions.R
\name{get_category_transitions}
\alias{get_category_transitions}
\alias{get_category_transitions.casal2MPD}
\alias{get_category_transitions.list}
\alias{get_category_transitions.casal2TAB}
\title{get_category_transitions}
\usage{
get_category_transitions(model, ...)

\method{get_category_transitions}{casal2MPD}(model, reformat_labels = TRUE)

\method{get_category_transitions}{list}(mode, reformat_labels = TRUE)

\method{get_category_transitions}{casal2TAB}(model, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of process type category_transitions
}
\author{
Craig Marsh
}
