% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_composition_observations.R
\name{get_composition_observations}
\alias{get_composition_observations}
\alias{get_composition_observations.casal2MPD}
\alias{get_composition_observations.list}
\alias{get_composition_observations.casal2TAB}
\title{get_composition_observations}
\usage{
get_composition_observations(model, ...)

\method{get_composition_observations}{casal2MPD}(model, reformat_labels = TRUE)

\method{get_composition_observations}{list}(model, reformat_labels = TRUE)

\method{get_composition_observations}{casal2TAB}(model, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
data frame with all observations of type == 'observation' and observation_type !%in% c('biomass', 'abundance')
}
\description{
An accessor function that returns a data frame of all composition data sets in a model
}
\author{
Craig Marsh
}
