% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariance.R
\name{get_covariance}
\alias{get_covariance}
\alias{get_covariance.casal2MPD}
\title{get_covariance}
\usage{
get_covariance(model)

\method{get_covariance}{casal2MPD}(model)
}
\arguments{
\item{model}{<casal2MPD, list> object that are generated from the extract.mpd() functions.}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a covariance function
}
\author{
Craig Marsh
}
