% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_derived_quanitites.R
\name{get_dqs}
\alias{get_dqs}
\title{get_dqs}
\usage{
get_dqs(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all derived quantitiy reports from Casal2 model output
}
\description{
get_dqs
}
