% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inital_partition.R
\name{get_initial_partition}
\alias{get_initial_partition}
\alias{get_initial_partition.casal2MPD}
\alias{get_initial_partition.list}
\title{get_initial_partition}
\usage{
get_initial_partition(model)

\method{get_initial_partition}{casal2MPD}(model)

\method{get_initial_partition}{list}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, list> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of process type initialisation_partition
}
\author{
Craig Marsh
}
