% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile.R
\name{get_profile}
\alias{get_profile}
\alias{get_profile.casal2MPD}
\title{get_profile}
\usage{
get_profile(model, ...)

\method{get_profile}{casal2MPD}(model, aggregate_obs = TRUE, reformat_labels = TRUE)
}
\arguments{
\item{model}{<casal2MPD> object that are generated from the extract.mpd() function}

\item{aggregate_obs}{bool, whether aggregate_obs observations log-likelihoods over all years}

\item{reformat_labels}{<bool> Reformat default Casal2 report labels to remove leading and trailing underscores (default = TRUE)}
}
\value{
A data frame with profile_values and likelihood components
}
\description{
An accessor function that returns a data frame of a profile class that can be easily plotted
}
\author{
Craig Marsh
}
