% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projections.R
\name{get_projections}
\alias{get_projections}
\alias{get_projections.casal2TAB}
\title{get_projections}
\usage{
get_projections(model)

\method{get_projections}{casal2TAB}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}
}
\value{
A data frame with all selectivity reports from Casal2 model output
}
\description{
An accessor function that returns a data frame from a Casal2 model output of projections
}
\author{
Craig Marsh
}
