% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compositional_observations.R
\name{plot_compositional_observations}
\alias{plot_compositional_observations}
\alias{plot_compositional_observations.casal2MPD}
\title{plot_compositional_observations}
\usage{
plot_compositional_observations(
  model,
  report_labels = NULL,
  plot_type = "classic",
  plot.it = T
)

\method{plot_compositional_observations}{casal2MPD}(
  model,
  report_labels = NULL,
  plot_type = "classic",
  plot.it = T
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions.}

\item{report_labels}{vector<string> report labels to plot, default is all}

\item{plot_type}{string
\itemize{
  \item classic
  \item residual
}}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}
}
\value{
A ggplot object

\code{NULL}
}
\description{
A plotting function for Casal2 fits to age and length compositional data
}
\examples{
\dontrun{

}
}
\author{
Craig Marsh
}
