% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\alias{plot_recruitment.casal2MPD}
\alias{plot_recruitment.casal2TAB}
\title{plot_recruitment plot recuitment}
\usage{
plot_recruitment(model, report_label = NULL, quantity = "ycs_values")

\method{plot_recruitment}{casal2MPD}(model, report_label = NULL, quantity = "ycs_values")

\method{plot_recruitment}{casal2TAB}(model, report_label = "", quantity = "ycs_values")
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions using the Casal2 base library}

\item{report_label}{<string> optional if you want to plot a specific recruitement report}

\item{quantity}{\itemize{
  \item ycs_values
  \item Recruits
  \item true_ycs
  \item standardised_ycs
}}
}
\value{
generate a plot over time uses get_BH_recruitment

\code{NULL}

\code{NULL}
}
\description{
A plotting function to plot recuitment for 'casal2TAB' and 'casal2MPD' objects.
}
\details{
If you have multiple time-steps and Recruitment
}
\author{
Craig Marsh
}
