% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_selectivities.R
\name{plot_selectivities}
\alias{plot_selectivities}
\title{plot_selectivities}
\usage{
plot_selectivities(model, selectivity_labels = NULL)
}
\arguments{
\item{model}{data.frame}

\item{selectivity_labels}{optional command if you only want to plot a subset of the available selectivities}
}
\value{
a ggplot
}
\description{
A Generic plotting function for selectivities that are derived from get_selectivities function()
}
\author{
Craig Marsh
}
