% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.simulated.data.R
\name{read.simulated.data}
\alias{read.simulated.data}
\title{read.simulated.data}
\usage{
read.simulated.data(dir, verbose = FALSE, mean_age = TRUE)
}
\arguments{
\item{dir}{path directory to a folder containing simulated observation from a casal2 -s run}

\item{verbose}{true print out statements throught function to help debug}

\item{mean_age}{<bool> if data is composition. True will return mean age/length for each year and data set. Otherwise False will return a list. each element will be a year which will contain a matrix of simulations x bin simulated values.}
}
\value{
a list of all the simulated data sets. in each element there is a matrix with rows = year, col = simulation reps
}
\description{
A utility function for reading in Casal2 simualted data sets. For help on simulated data from Casal2 see
}
\author{
Craig Marsh
}
