% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_report_label.R
\name{reformat_default_labels}
\alias{reformat_default_labels}
\title{Utility function to reformat report labels created from
Casal2s default report. cut off leading and trailing __
so the report label '__selectivities__' will be changed to 'selectivities'
if a default report has been created with the same label as an exiting report
we change its name from '__report_label__' to 'report_label_default'}
\usage{
reformat_default_labels(report_labels)
}
\arguments{
\item{report_labels}{vector of report labels}
}
\description{
Utility function to reformat report labels created from
Casal2s default report. cut off leading and trailing __
so the report label '__selectivities__' will be changed to 'selectivities'
if a default report has been created with the same label as an exiting report
we change its name from '__report_label__' to 'report_label_default'
}
\author{
Craig Marsh
}
\keyword{internal}
