% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{StripComments}
\alias{StripComments}
\title{StripComments}
\usage{
StripComments(file)
}
\arguments{
\item{file}{a file read in by scan}
}
\value{
a file that has removed lines starting with # or betwee '/*' '*/'
}
\description{
A utility function for stripping out comments that is
line start with #
or a wrapped in /* */
}
\author{
Craig Marsh
}
