% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_config.R
\name{summarise_config}
\alias{summarise_config}
\title{summarise_config}
\usage{
summarise_config(
  config_dir = "",
  config_file = "config.csl2",
  quiet = T,
  fileEncoding = ""
)
}
\arguments{
\item{config_dir}{path directory}

\item{config_file}{the starting config that starts a Casal2 model default config.csl2, but can be overridden with casal2 -c my_config. in the later case it should be the my_config name}
}
\value{

}
\description{
utility function that reads in config.csl2 file and summarises a model in table formats and plots
}
\examples{
\dontrun{
library(r4Casal2)
library(ggplot2)
config_dir = system.file("extdata", "TestModelComplex", package = "r4Casal2", mustWork = TRUE)
summary = summarise_config(config_dir, config_file = "config.csl2", quiet = T)
names(summary)
# visualise observations in model
ggplot(summary$obs_year_df, aes(x = year, y = observation, col = observation, size = active)) +
geom_point() +
guides(colour = "none", size = "none")
}
}
\author{
Craig Marsh
}
