% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_estimated_parameters.R
\name{summarise_estimated_parameters}
\alias{summarise_estimated_parameters}
\alias{summarise_estimated_parameters.casal2MPD}
\title{summarise_estimated_parameters}
\usage{
summarise_estimated_parameters(model, ...)

\method{summarise_estimated_parameters}{casal2MPD}(model, plot_it = FALSE, ignore = NULL, n = 100)
}
\arguments{
\item{model}{an MPD model}

\item{plot_it}{if true plot MPD value, initial value, and prior, otherwise report MPD, init, lower_bound, upper_bound, prior_type}

\item{ignore}{ignore parameters matching a string (e.g., "YCS_values")}

\item{n}{Number of values to evaluate so as provide enough points to plot a smooth line (default = 100)}
}
\value{
either a ggplot or a list with two elements. The first is for plotting, the second is a useful data frame
outlining priors, bounds mpd values initial values etc.
}
\description{
this function will find estimate_summary report and summarise the priors, MPD and initial value
}
\author{
C Marsh
}
